новый метод репозитория saveAll для массовой вставки, тест не проходит

+++ src/main/java/ru/inno/adeliya/jdbc/repository/generator/SequenceWithBatchesGenerator.java
@@ -7,10 +7,10 @@ import java.sql.Statement;
 import java.util.LinkedList;
 import java.util.Queue;
 
-public class SequenceWithBatchesGenerator implements IdGenerator<Long> {
+public class SequenceWithBatchesGenerator implements IdGenerator<Integer> {
     private final Connection connection;
     private final int batchSize;
-    private final Queue<Long> idQueue = new LinkedList<>();
+    private final Queue<Integer> idQueue = new LinkedList<>();
     private final String sequenceName;
 
     public SequenceWithBatchesGenerator(Connection connection, int batchSize, String sequenceName) {
@@ -20,13 +20,13 @@ public class SequenceWithBatchesGenerator implements IdGenerator<Long> {
     }
 
     @Override
-    public synchronized Long generate() {
+    public synchronized Integer generate() {
         if (idQueue.isEmpty()) {
             String query = String.format("SELECT nextval('%s') FROM generate_series(1, %d)", sequenceName, batchSize);
             try (Statement statement = connection.createStatement();
                  ResultSet resultSet = statement.executeQuery(query)) {
                 while (resultSet.next()) {
-                    idQueue.offer(resultSet.getLong(1));
+                    idQueue.offer(resultSet.getInt(1));
                 }
             } catch (SQLException e) {
                 throw new RuntimeException(e);
